+++ NEWS.md
@@ -5,6 +5,8 @@
 
 #### Highlights
 * Fix bug in checking of PlotMapPoints() and PlotMapOutput() legend positions
+* Fix bug in PlotMapPoints() where custom color breaks were ignored
+* Add option to reverse color palette in PlotMapPoints()
 
 ## HYPEtools 1.3.0 (2023-04-05)
 *New functions and bug fixes*